--- libs/shared/ui/src/lib/multi-step-form/index.ts
@@ -1,3 +1,0 @@
-export * from './multi-step-form-datasource';
-export * from './multi-step-form.component';
-export * from './form-step-component';
