+++ libs/simulation-project-utils/ui/src/lib/create-project/create-project/create-project.component.html
@@ -6,7 +6,7 @@
     a SED-ML file generated from the form below. You can then use runBioSimulations to execute your archive, or you can
     download the archive.
   </div>
-  <multi-step-form [dataSource]="formDataSource"></multi-step-form>
+  <paging-form [dataSource]="formDataSource"></paging-form>
 </biosimulations-page>
 
 <ng-template #loading>
